UPDATE: funcionairoService.js
Adaptando o service de funcionario para utilizar o campo eADM

+++ backend/services/funcionarioService.js
@@ -1,10 +1,11 @@
+const DataError = require('../models/DataError')
 const Funcionario = require('../models/Funcionario')
 
 
 exports.createFuncionario = async (data) => {
     try {
-        const {cpf, nome, dataNascimento, telefone, email, senha} = data
-        const novoFuncionario = new Funcionario(nome, cpf, dataNascimento, telefone, email, senha)
+        const {cpf, nome, dataNascimento, telefone, email, senha, eADM} = data
+        const novoFuncionario = new Funcionario(nome, cpf, dataNascimento, telefone, email, senha, eADM)
 
         await novoFuncionario.create()
 
@@ -40,7 +41,7 @@ exports.getFuncionarioById = async (id) => {
 
 exports.updateFuncionario = async (data) => {
     try {
-        const {id, cpf, nome, dataNascimento, telefone, email, senha} = data
+        const {id, cpf, nome, dataNascimento, telefone, email, senha, eADM} = data
         const funcionario = await Funcionario.findById(id)
 
         if(!funcionario){
@@ -53,6 +54,7 @@ exports.updateFuncionario = async (data) => {
         funcionario.telefone = telefone
         funcionario.email = email
         funcionario.senha = senha
+        funcionario.eADM = eADM
 
         await funcionario.update()
 
@@ -62,12 +64,28 @@ exports.updateFuncionario = async (data) => {
     }
 }
 
-exports.deleteFuncionario = async (id) => {
+exports.deleteFuncionario = async (data) => {
     try {
-        const funcionario = await Funcionario.findById(id)
+        const {idAdm, idDelete} = data
+
+        const adm = await Funcionario.findById(idAdm)
+
+        if(!adm){
+            throw new DataError('Not Found', 404, 'ADM não encontrado')
+        }
+
+        if(!adm.eADM){
+            throw new DataError('Permission Error', 403, 'Funcionario não é ADM para efetuar a ação')
+        }
+
+        const funcionario = await Funcionario.findById(idDelete)
 
         if(!funcionario){
-            throw new DataError('Not Found', 404, 'Funcionario não encontrado')
+            throw new DataError('Not Found', 404, 'Funcionário não encontrado')
+        }
+
+        if(adm.id === funcionario.id){
+            throw new DataError('Permission Erro', 403, 'ADM não pode se deletar do sistema')
         }
 
         await funcionario.delete()
